Fix remaining size_t vs int mismatch in GeoRSS and WaSP
git-svn-id: https://svn.osgeo.org/gdal/trunk@31626 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/georss/ogrgeorsslayer.cpp
@@ -960,7 +960,7 @@ static int OGRGeoRSSLayerIsStandardFieldInternal(const char* pszName,
             size_t nLen = strlen(papszNames[i]);
             if (strncmp(pszName, papszNames[i], nLen) == 0)
             {
-                int k = nLen;
+                size_t k = nLen;
                 while(pszName[k] >= '0' && pszName[k] <= '9')
                     k++;
                 if (pszName[k] == '\0')
